intermediate commit of DAC code refactoring. DAC output currently broken
